+++ .packages.yaml
@@ -1,6 +1,6 @@
 '@ssen/*':
-  version: 0.2.2
+  version: 0.2.4
   tag: canary
-rocket-pack:
-  version: 0.9.0-alpha.2
+rocket-punch:
+  version: 0.9.0-alpha.4
   tag: canary
