+++ src/rocket-punch/packageJson/getExternalPackages.ts
