+++ src/rocket-punch/packageJson/getInternalPackages.ts
