+++ src/rocket-punch/rule/fsCopySourceFilter.ts
