+++ src/rocket-punch/rule/readDirectoryPatterns.ts
